# apps/sgd-next-console/src/types/openapi.json
{
  "openapi": "3.1.0",
  "components": {
    "schemas": {
...
      "CategoryCreateRequest": {
        "type": "object",
        "required": [
          "sportId",
          "name",
          "minAge",
          "maxAge"
        ],
        "properties": {
          "sportId": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "format": "int64"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 255,
            "pattern": "\\S"
          },
          "minAge": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 0
          },
          "maxAge": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 0
          },
          "ageRangeValid": {
            "type": "boolean"
          },
          "ageRange": {
            "type": "string"
          }
        }
      },
...
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "minAge": {
            "type": "integer",
            "format": "int32"
          },
          "maxAge": {
            "type": "integer",
            "format": "int32"
          },
          "ageRange": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "sport": {
            "$ref": "#/components/schemas/SportInfo1"
          },
          "createdAt": {
            "$ref": "#/components/schemas/LocalDateTime"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/LocalDateTime"
          }
        }
      },
      "CategoryUpdateRequest": {
        "type": "object",
        "required": [
          "sportId",
          "name",
          "minAge",
          "maxAge"
        ],
        "properties": {
          "sportId": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "format": "int64"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 255,
            "pattern": "\\S"
          },
          "minAge": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 0
          },
          "maxAge": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 0
          },
          "ageRangeValid": {
            "type": "boolean"
          },
          "ageRange": {
            "type": "string"
          }
        }
      },
...
      "LocalDateTime": {
        "type": "string",
        "format": "date-time",
        "examples": [
          "2022-03-10T12:15:50"
        ]
      },
...
      "SportInfo1": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT token obtenido a través de Keycloak. El token debe incluir los roles necesarios (ADMIN_GENERAL, ADMIN_CLUB, PROFESOR) y los claims personalizados (club_id, venue_ids, sport_ids).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "SecurityScheme": {
        "type": "openIdConnect",
        "openIdConnectUrl": "http://localhost:8089/realms/SGD/.well-known/openid-configuration",
        "description": "Authentication"
      }
    }
  },
  "info": {
    "contact": {
      "email": "dev@sgd.com",
      "name": "SGD Development Team",
      "url": "https://sgd.com/support"
    },
    "description": "Sistema de Gesti�n Deportiva - API Backend",
    "license": {
      "name": "Proprietary",
      "url": "https://sgd.com/license"
    },
    "title": "SGD Backend API",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.sgd.com",
      "description": "Servidor de producción"
    }
  ],
...
  "paths": {
...
    "/api/categories": {
      "get": {
        "summary": "Obtener categorías",
        "description": "Lista categorías activas con filtros opcionales por deporte y/o edad.",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "description": "Edad para filtrar categorías válidas",
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 0
            },
            "name": "age",
            "in": "query"
          },
          {
            "description": "ID del deporte para filtrar",
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "name": "sportId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de categorías recuperada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parámetros de consulta inválidos"
          },
          "401": {
            "description": "Token de acceso inválido o expirado"
          },
          "403": {
            "description": "Sin permisos para acceder a categorías"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Crear nueva categoría",
        "description": "Crea una nueva categoría deportiva. Solo disponible para ADMIN_GENERAL.",
        "tags": [
          "Categories"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Categoría creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos"
          },
          "401": {
            "description": "Token de acceso inválido o expirado"
          },
          "403": {
            "description": "Sin permisos para crear categorías"
          },
          "404": {
            "description": "Deporte no encontrado"
          },
          "409": {
            "description": "Rango de edad se superpone con categoría existente"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/categories/by-age/{age}": {
      "get": {
        "summary": "Obtener categorías por edad",
        "description": "Lista todas las categorías válidas para una edad específica en todos los deportes.",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "description": "Edad del atleta",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 0
            },
            "name": "age",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de categorías válidas para la edad",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Edad inválida"
          },
          "401": {
            "description": "Token de acceso inválido"
          },
          "403": {
            "description": "Sin permisos para acceder a categorías"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/categories/{id}": {
      "put": {
        "summary": "Actualizar categoría existente",
        "description": "Actualiza los datos de una categoría existente. Solo disponible para ADMIN_GENERAL.",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "description": "ID único de la categoría",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Categoría actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos"
          },
          "401": {
            "description": "Token de acceso inválido o expirado"
          },
          "403": {
            "description": "Sin permisos para actualizar categorías"
          },
          "404": {
            "description": "Categoría o deporte no encontrado"
          },
          "409": {
            "description": "Rango de edad se superpone con categoría existente"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Obtener categoría específica",
        "description": "Obtiene los detalles completos de una categoría incluyendo información del deporte.",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "description": "ID único de la categoría",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Categoría encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "403": {
            "description": "Sin permisos para ver esta categoría"
          },
          "404": {
            "description": "Categoría no encontrada"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Eliminar categoría",
        "description": "Elimina una categoría del sistema (soft delete). Solo disponible para ADMIN_GENERAL.",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "description": "ID único de la categoría",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Categoría eliminada exitosamente"
          },
          "400": {
            "description": "No se puede eliminar la categoría (tiene atletas asociados)"
          },
          "401": {
            "description": "Token de acceso inválido o expirado"
          },
          "403": {
            "description": "Sin permisos para eliminar categorías"
          },
          "404": {
            "description": "Categoría no encontrada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}